app/requirements.txt
flask==3.1.0
python3-saml==1.16.0
lxml
xmlsec
python-dotenv
